@@ -5,11 +5,11 @@
   "homepage": "https://www.mta-trader.com",
   "dependencies": {
     "bootstrap": "^5.1.3",
-    "gh-pages": "^3.2.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-scripts": "5.0.0",
-    "reactstrap": "^9.0.1"
+    "gh-pages": "^4.0.0",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
+    "react-scripts": "5.0.1",
+    "reactstrap": "^9.1.1"
   },
   "scripts": {
     "start": "react-scripts start",
